Take and select a photo features were added
The user can now take a photo from the camera app, decide if save it or
no.
The saved pictures have now a format (yyyyMMddHHmmss)for the file name.
A functionality to select a photo from the image gallery was added.
The action bar was modified. There were added three buttons. From left
to right; the first one is to share the content (still need to be
programmed), the second one is to takeselect a picture(already
programmed), the third one is for the settings(still need to be
programmed and may be taken out of the app, I just added it to have a
settings button inside the app and not only in the device's buttons
menu, but it feels useless since I don't think there would be options at
all). BTW the icons for those buttons were added int their respective
mipmap folders.
The strings.xml file was also complemented with some new strings (Not a
big deal for the guy that will translate this ;D ).
Also the ipAddress activity changed since the previous version of the
app. Now it is not the first activity shown, it will only be shown if
and only if the user will send the pictures via UDP and he has not
specified yet the ip address and the port.
BrowseImageActivity may be deleted in the next commit, at this point of
the project, I don't feel like it will have any use. Instead I think it
will be better to use only the canvas activity.
The sender still need to be fixed. Sincerely, this part is a pain in the
***.